Create endpoint Implementation
The following code shows how the /estimate endpoint can be implemented in Python:

This code will connect to a database of cost estimates and return the cost estimates for the user's location, income, and family size. The cost estimates can be returned in JSON format, which can be easily parsed by the front end.