Create CostEstimates front end
The following code shows how the front end can be updated to display the cost estimates:

This code will display the cost estimates for each category in a list. The list can be easily customized to display the cost estimates in a way that is easy for the user to understand.